+++ tests/features/environment.py
@@ -7,6 +7,8 @@ import sys
 import db_utils as dbutils
 import fixture_utils as fixutils
 import pexpect
+import tempfile
+import shutil
 
 from steps.wrappers import run_cli, wait_prompt
 
@@ -96,6 +98,10 @@ def before_all(context):
 
     context.fixture_data = fixutils.read_fixture_files()
 
+    context.env_config_home = tempfile.mkdtemp(prefix='pgcli_home_')
+
+    os.environ['XDG_CONFIG_HOME'] = context.env_config_home
+
 
 def after_all(context):
     """
@@ -105,6 +111,9 @@ def after_all(context):
     dbutils.drop_db(context.conf['host'], context.conf['user'],
                     context.conf['pass'], context.conf['dbname'],
                     context.conf['port'])
+
+    # Remove temp config direcotry
+    shutil.rmtree(context.env_config_home)
 
     # Restore env vars.
     for k, v in context.pgenv.items():